Polishing #1122
Rename wrapper type to ComposedRedisCodec. Introduce RedisCodec.of(…) factory method.

Add tests.

Original pull request: #1123.
